# docs/swagger.yaml
basePath: /v1
definitions:
  health_worker.HTTPError:
    type: object
  model.HealthCheck:
    properties:
      check_interval:
        type: integer
      id:
        type: integer
      name:
        type: string
      params:
        $ref: '#/definitions/model.HealthCheckParams'
        type: object
      routing_polilcies:
        $ref: '#/definitions/model.RoutingPolicies'
        type: object
      threshould:
        type: integer
      type:
        type: integer
    type: object
  model.HealthCheckParams:
    properties:
      addr:
        type: string
      hostName:
        type: string
      path:
        type: string
      port:
        type: integer
      protocol:
        type: integer
      searchWord:
        type: string
      timeout:
        type: integer
    type: object
  model.RoutingPolicies:
    items:
      properties:
        health_check_id:
          type: integer
        id:
          type: integer
        record_id:
          type: integer
        type:
          type: integer
      type: object
    type: array
  model.RoutingPolicy:
    properties:
      health_check_id:
        type: integer
      id:
        type: integer
      record_id:
        type: integer
      type:
        type: integer
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: This is PIR5 HealthCheck worker and API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: health-worker
  version: "1.0"
paths:
  /healthchecks:
    get:
      consumes:
      - application/json
      description: get healthchecks
      parameters:
      - description: HealthCheck ID
        in: query
        name: id
        type: integer
      - description: Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.HealthCheck'
            type: array
        "404":
          description: Not Found
          schema:
            items:
              $ref: '#/definitions/model.HealthCheck'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: get healthchecks
    post:
      consumes:
      - application/json
      description: create healthCheck
      parameters:
      - description: HealthCheck Object
        in: body
        name: healthCheck
        required: true
        schema:
          $ref: '#/definitions/model.HealthCheck'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HealthCheck'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: create healthCheck
  /healthchecks/{id}:
    delete:
      consumes:
      - application/json
      description: delete healthCheck
      parameters:
      - description: HealthCheck ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.HealthCheck'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: delete healthCheck
    put:
      consumes:
      - application/json
      description: update healthCheck
      parameters:
      - description: HealthCheck ID
        in: path
        name: id
        required: true
        type: integer
      - description: HealthCheck Object
        in: body
        name: healthCheck
        required: true
        schema:
          $ref: '#/definitions/model.HealthCheck'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthCheck'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: update healthCheck
  /routingpolicies:
    get:
      consumes:
      - application/json
      description: get routingpolicies
      parameters:
      - description: RoutingPolicy ID
        in: query
        name: id
        type: integer
      - description: Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.RoutingPolicy'
            type: array
        "404":
          description: Not Found
          schema:
            items:
              $ref: '#/definitions/model.RoutingPolicy'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: get routingpolicies
    post:
      consumes:
      - application/json
      description: create healthCheck
      parameters:
      - description: RoutingPolicy Object
        in: body
        name: healthCheck
        required: true
        schema:
          $ref: '#/definitions/model.RoutingPolicy'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RoutingPolicy'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: create healthCheck
  /routingpolicies/{id}:
    delete:
      consumes:
      - application/json
      description: delete healthCheck
      parameters:
      - description: RoutingPolicy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.RoutingPolicy'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: delete healthCheck
    put:
      consumes:
      - application/json
      description: update healthCheck
      parameters:
      - description: RoutingPolicy ID
        in: path
        name: id
        required: true
        type: integer
      - description: RoutingPolicy Object
        in: body
        name: healthCheck
        required: true
        schema:
          $ref: '#/definitions/model.RoutingPolicy'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RoutingPolicy'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/health_worker.HTTPError'
            type: object
      security:
      - ID: []
      - Secret: []
      summary: update healthCheck
  /viron:
    get:
      consumes:
      - application/json
      description: get global menu
      operationId: viron#get
      produces:
      - application/json
      summary: get global menu
      tags:
      - viron
  /viron_authtype:
    get:
      consumes:
      - application/json
      description: get auth type
      operationId: viron_authtype#get
      produces:
      - application/json
      summary: get auth type
      tags:
      - viron
securityDefinitions:
  Bearer:
    in: header
    name: Bearer
    type: apiKey
swagger: "2.0"